Move getDefaultName and getDefaultEnumValues to wildcard.ts
